[VQueues] Move scheduler status types to worker-api
This is to allow datafusion storage engine to access these types without direct dependency on restate-vqueues.
